refactor querying Fixes #35
