+++ cmd/info.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/rs/zerolog/log"
-	"github.com/spagettikod/vsix/vscode"
+	"github.com/spagettikod/vsix/marketplace"
 
 	"github.com/spf13/cobra"
 )
@@ -23,7 +23,7 @@ var infoCmd = &cobra.Command{
 	DisableFlagsInUseLine: true,
 	Run: func(cmd *cobra.Command, args []string) {
 		log.Info().Str("identifier", args[0]).Msg("looking up extension at Marketplace")
-		ext, err := vscode.NewExtension(args[0])
+		ext, err := marketplace.FetchExtension(args[0])
 		if err != nil {
 			fmt.Println(err)
 			os.Exit(1)
