+++ cmd/versions.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/rs/zerolog/log"
-	"github.com/spagettikod/vsix/vscode"
+	"github.com/spagettikod/vsix/marketplace"
 
 	"github.com/spf13/cobra"
 )
@@ -22,7 +22,7 @@ var versionsCmd = &cobra.Command{
 	DisableFlagsInUseLine: true,
 	Run: func(cmd *cobra.Command, args []string) {
 		log.Info().Str("identifier", args[0]).Msg("looking up extension at Marketplace")
-		ext, err := vscode.NewExtension(args[0])
+		ext, err := marketplace.FetchExtension(args[0])
 		if err != nil {
 			fmt.Println(err)
 			os.Exit(1)
