+++ marketplace/query.go
@@ -1,12 +1,20 @@
-package vscode
+package marketplace
 
 import (
 	"encoding/json"
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"os"
+	"strings"
+
+	"github.com/spagettikod/vsix/vscode"
 )
 
 type QueryResults struct {
 	Results []struct {
-		Extensions     []Extension `json:"extensions"`
+		Extensions     []vscode.Extension `json:"extensions"`
 		ResultMetadata []struct {
 			MetadataType  string `json:"metadataType"`
 			MetadataItems []struct {
@@ -57,11 +65,15 @@ const (
 	FilterTypeTarget        FilterType = 8
 	FilterTypeFeatured      FilterType = 9
 	FilterTypeSearchText    FilterType = 10
+
+	debugEnvVar = "VSIX_DEBUG"
 )
 
 var (
-	MSVSCodeCriteria    = Criteria{FilterType: FilterTypeTarget, Value: "Microsoft.VisualStudio.Code"}
-	SomeUnknownCriteria = Criteria{FilterType: 12, Value: "4096"}
+	MSVSCodeCriteria          = Criteria{FilterType: FilterTypeTarget, Value: "Microsoft.VisualStudio.Code"}
+	SomeUnknownCriteria       = Criteria{FilterType: 12, Value: "4096"}
+	ErrExtensionNotFound      = errors.New("extension could not be found at Marketplace")
+	ErrExtensionHasNoVersions = errors.New("extension has no versions")
 )
 
 var latestVersionQueryTemplate2 = Filter{
@@ -69,6 +81,29 @@ var latestVersionQueryTemplate2 = Filter{
 		MSVSCodeCriteria,
 		SomeUnknownCriteria,
 	},
+}
+
+type extensionQueryResponse struct {
+	Results []struct {
+		Extensions []vscode.Extension `json:"extensions"`
+	} `json:"results"`
+}
+
+func NewQuery() Query {
+	q := Query{}
+	f := Filter{
+		Criteria: []Criteria{
+			MSVSCodeCriteria,
+			SomeUnknownCriteria,
+		},
+		PageNumber: 1,
+		PageSize:   1,
+		SortBy:     ByNone,
+		SortOrder:  0,
+	}
+	q.Filters = append(q.Filters, f)
+	q.Flags = FlagLatestVersion
+	return q
 }
 
 func (q Query) AddCriteria(c Criteria) {
@@ -93,46 +128,51 @@ func (q Query) CriteriaValues(filterType FilterType) []string {
 	return values
 }
 
-func NewQuery() Query {
-	q := Query{}
-	f := Filter{
-		Criteria: []Criteria{
-			MSVSCodeCriteria,
-			SomeUnknownCriteria,
-		},
-		PageNumber: 1,
-		PageSize:   1,
-		SortBy:     ByNone,
-		SortOrder:  0,
+func (q Query) Run() (extensionQueryResponse, error) {
+	if _, debug := os.LookupEnv(debugEnvVar); debug {
+		ioutil.WriteFile("query.json", []byte(q.ToJSON()), 0644)
 	}
-	q.Filters = append(q.Filters, f)
-	q.Flags = FlagLatestVersion
-	return q
+	eqr := extensionQueryResponse{}
+	req, err := http.NewRequest(http.MethodPost, "https://marketplace.visualstudio.com/_apis/public/gallery/extensionquery", strings.NewReader(q.ToJSON()))
+	if err != nil {
+		return eqr, err
+	}
+	req.Header.Add("Content-Type", "application/json")
+	req.Header.Add("Accept", "application/json;api-version=3.0-preview.1")
+	c := http.Client{}
+	resp, err := c.Do(req)
+	if err != nil {
+		return eqr, err
+	}
+	if resp.StatusCode != http.StatusOK {
+		return eqr, fmt.Errorf("marketplace.visualstudio.com returned HTTP %v", resp.StatusCode)
+	}
+	b, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return eqr, err
+	}
+	if _, debug := os.LookupEnv(debugEnvVar); debug {
+		ioutil.WriteFile("response.json", b, 0644)
+	}
+	err = json.Unmarshal(b, &eqr)
+	if err != nil {
+		return eqr, err
+	}
+	if len(eqr.Results[0].Extensions) == 0 {
+		return eqr, ErrExtensionNotFound
+	}
+	if len(eqr.Results[0].Extensions[0].Versions) == 0 {
+		return eqr, ErrExtensionHasNoVersions
+	}
+	return eqr, err
 }
 
-func LatestQueryJSON(uniqueID string) string {
-	q := NewQuery()
-	q.AddCriteria(Criteria{
-		FilterType: FilterTypeExtensionName,
-		Value:      uniqueID,
-	})
-	q.Flags = FlagLatestVersion
-	b, _ := json.Marshal(q)
-	return string(b)
+func (q Query) ToJSON() string {
+	qjson, _ := json.Marshal(q)
+	return string(qjson)
 }
 
-func listVersionsQueryJSON(uniqueID string) string {
-	q := NewQuery()
-	q.AddCriteria(Criteria{
-		FilterType: FilterTypeExtensionID,
-		Value:      uniqueID,
-	})
-	q.Flags = FlagAllVersions
-	b, _ := json.Marshal(q)
-	return string(b)
-}
-
-func searchQueryJSON(query string, limit int, sortBy SortCriteria) string {
+func QueryLastestVersionByText(query string, limit int, sortBy SortCriteria) Query {
 	q := NewQuery()
 	q.AddCriteria(Criteria{
 		FilterType: FilterTypeSearchText,
@@ -141,6 +181,25 @@ func searchQueryJSON(query string, limit int, sortBy SortCriteria) string {
 	q.Flags = FlagLatestVersion
 	q.Filters[0].SortBy = sortBy
 	q.Filters[0].PageSize = limit
-	b, _ := json.Marshal(q)
-	return string(b)
+	return q
+}
+
+func QueryLatestVersionByUniqueID(uniqueID string) Query {
+	q := NewQuery()
+	q.AddCriteria(Criteria{
+		FilterType: FilterTypeExtensionName,
+		Value:      uniqueID,
+	})
+	q.Flags = FlagLatestVersion
+	return q
+}
+
+func QueryAllVersionsByUniqueID(uniqueID string) Query {
+	q := NewQuery()
+	q.AddCriteria(Criteria{
+		FilterType: FilterTypeExtensionID,
+		Value:      uniqueID,
+	})
+	q.Flags = FlagAllVersions
+	return q
 }
